requirements.txt
pypdf2~=3.0.1
pdfminer.six